backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic[email]
pydantic-settings
email-validator
python-multipart
python-telegram-bot
ollama
httpx
python-jose[cryptography]
passlib[bcrypt]
pillow
aiosqlite
python-dotenv
